+++ src/gui/desktop/quick_views/CMakeLists.txt
@@ -11,8 +11,8 @@ qt6_add_resources(QML_RESOURCES
 
 qt6_wrap_cpp(MOCED
                 amedia_item.hpp
-                bridge.hpp
                 media_view_item.hpp
+                objects_accessor.hpp
                 photo_data_qml.hpp
                 photo_item.hpp
                 photos_model_controller_component.hpp
@@ -26,10 +26,10 @@ add_library(quick_views OBJECT
                 ${MOCED}
                 amedia_item.cpp
                 amedia_item.hpp
-                bridge.cpp
-                bridge.hpp
                 media_view_item.cpp
                 media_view_item.hpp
+                objects_accessor.cpp
+                objects_accessor.hpp
                 photo_data_qml.cpp
                 photo_data_qml.hpp
                 photo_item.cpp
