+++ src/gui/desktop/quick_views/objects_accessor.cpp
@@ -1,5 +1,5 @@
 
-#include "bridge.hpp"
+#include "objects_accessor.hpp"
 
 
 ObjectsAccessor::ObjectsAccessor(QObject* parent)
