+++ src/gui/desktop/quick_views/objects_accessor.hpp
